refactor rules list
